moved chat components to folder
