+++ client/src/components/Chat/ChatFriendsListItem.js
