+++ client/src/components/Chat/ChatFriendsSideBar.js
