+++ client/src/pages/ChatPage.js
@@ -1,10 +1,11 @@
-import React, {useEffect} from 'react';
+import React, {useEffect, useState} from 'react';
 import { styled } from "@mui/system";
-import ChatSideBar from '../components/ChatSideBar'
-import ChatFriendsSideBar from '../components/ChatFriendsSideBar';
-import ChatAppBar from '../components/ChatAppBar';
-import Messenger from '../components/Messenger';
+import ChatSideBar from '../components/Chat/ChatSideBar'
+import ChatFriendsSideBar from '../components/Chat/ChatFriendsSideBar';
+import ChatAppBar from '../components/Chat/ChatAppBar';
+import Messenger from '../components/Chat/Messenger';
 import logout from '../utils/auth'
+import connectWithSocketServer from '../realtimeCommunication/socketConnection'
 
 const Wrapper = styled('div')({
   width: '100%',
@@ -13,13 +14,14 @@ const Wrapper = styled('div')({
 });
 
 export default function ChatPage() {
+  // const [userDetails, setUserDetails] = useState()
   useEffect(() => {
-    const userDetails = localStorage.getItem('id_token');
+    const userDetails = localStorage.getItem('email');
 
     if(!userDetails){
       logout();
-    // }else{
-    //   setUserDetails(JSON.parse(userDetails))
+    }else{
+      connectWithSocketServer(userDetails)
     }
   })
   return (
